+++ src/main/java/com/commerce/backend/service/cache/ItemObjectCategoryCacheService.java
@@ -22,5 +22,5 @@ public interface ItemObjectCategoryCacheService {
     BasicResponse createItemObjectCategory(CategoryRequest request);
     ItemObjectCategory deleteItemObjectCategory(Long id);
     ItemObjectCategory updateItemCategory(CategoryUpdateRequest request);
-    List<ItemObjectCategory> findCategoriesByParent(Integer id);
+    List<ItemObjectCategory> findCategoriesByParent(Long id);
 }
